+++ src/portal/src/app/base/global-search/search-results.ts
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 import { Project } from "../../project/project";
-import { Repository, HelmChartSearchResultItem } from "@harbor/ui";
+import { Repository } from "@harbor/ui";
+import { HelmChartSearchResultItem } from "../../project/helm-chart/helm-chart.interface.service";
 
 export class SearchResults {
   constructor() {
