+++ src/portal/src/app/project/helm-chart/helm-chart-detail/chart-detail.component.html
