+++ src/portal/src/app/project/helm-chart/helm-chart-detail/chart-detail.component.ts
@@ -1,9 +1,9 @@
-import { RoleMapping } from './../../shared/shared.const';
+import { RoleMapping } from './../../../shared/shared.const';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Component, OnInit } from "@angular/core";
-import { Project } from '../project';
-import { SessionService } from './../../shared/session.service';
-import { SessionUser } from './../../shared/session-user';
+import { Project } from '../../project';
+import { SessionService } from './../../../shared/session.service';
+import { SessionUser } from './../../../shared/session-user';
 import {finalize} from "rxjs/operators";
 
 @Component({
@@ -11,7 +11,7 @@ import {finalize} from "rxjs/operators";
   templateUrl: "./chart-detail.component.html",
   styleUrls: ["./chart-detail.component.scss"]
 })
-export class ChartDetailComponent implements OnInit {
+export class HelmChartDetailComponent implements OnInit {
 
   projectId: number | string;
   project: Project;
