+++ src/portal/src/app/project/helm-chart/helm-chart-detail/chart-detail/chart-detail.component.scss
@@ -1,4 +1,4 @@
-@import "../../mixin";
+@import "./../../../../../../lib/src/mixin";
 
 .title-container {
     display: flex;
