+++ src/portal/src/app/project/helm-chart/label-filter/label-filter.component.scss
