+++ src/portal/src/app/project/helm-chart/label-filter/label-filter.component.ts
@@ -3,9 +3,9 @@ import {ClrDatagridFilterInterface} from "@clr/angular";
 import { fromEvent } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
 
-import { Label, Tag } from './../../service/interface';
-import { HelmChartVersion } from '../../service/interface';
-import { ResourceType } from '../../shared/shared.const';
+import { Label, Tag } from '@harbor/ui';
+import { HelmChartVersion } from '../helm-chart.interface.service';
+import { ResourceType } from '../../../shared/shared.const';
 
 @Component({
     selector: "hbr-chart-version-label-filter",
