+++ src/portal/src/app/project/helm-chart/label-marker/label-marker.component.html
