+++ src/portal/src/app/project/helm-chart/list-chart-versions/helm-chart-versions-detail/helm-chart-version.component.scss
@@ -1,4 +1,4 @@
-@import "../../mixin";
+@import "../../../../../../lib/src/mixin";
 
 .title-container {
     display: flex;
