+++ src/portal/src/app/project/helm-chart/list-chart-versions/list-chart-versions.component.scss
