+++ src/portal/src/app/project/helm-chart/list-chart-versions/list-chart-versions.component.ts
@@ -2,9 +2,9 @@ import { Router } from '@angular/router';
 import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
 
-import { Project } from './../project';
-import { SessionUser } from './../../shared/session-user';
-import { SessionService } from './../../shared/session.service';
+import { Project } from './../../project';
+import { SessionUser } from './../../../shared/session-user';
+import { SessionService } from './../../../shared/session.service';
 
 @Component({
   selector: 'list-chart-version',
