+++ src/portal/src/app/project/helm-chart/list-charts-detail/helm-chart.component.html
