+++ src/portal/src/app/project/helm-chart/list-charts-detail/helm-chart.component.scss
@@ -1,4 +1,4 @@
-@import "../mixin";
+@import "../../../../../lib/src/mixin";
 $size24:24px;
 $size60:60px;
